Remove legacy AccessControlList mechanism
